refactor: Testcontainers 공통 설정을 IntegrationTestSupport로 분리
