+++ src/test/java/com/terning/farewell_server/FarewellServerApplicationTests.java
@@ -1,36 +1,9 @@
 package com.terning.farewell_server;
 
+import com.terning.farewell_server.util.IntegrationTestSupport;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.containers.KafkaContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
-@ActiveProfiles("test")
-@Testcontainers
-@SpringBootTest
-class FarewellServerApplicationTests {
-
-	@Container
-	private static final KafkaContainer KAFKA_CONTAINER =
-			new KafkaContainer(DockerImageName.parse("confluentinc/cp-kafka:7.5.3"));
-
-	@Container
-	private static final GenericContainer<?> REDIS_CONTAINER =
-			new GenericContainer<>(DockerImageName.parse("redis:7-alpine")).withExposedPorts(6379);
-
-	@DynamicPropertySource
-	static void properties(DynamicPropertyRegistry registry) {
-		registry.add("spring.data.redis.host", REDIS_CONTAINER::getHost);
-		registry.add("spring.data.redis.port", () -> REDIS_CONTAINER.getMappedPort(6379).toString());
-
-		registry.add("spring.kafka.bootstrap-servers", KAFKA_CONTAINER::getBootstrapServers);
-	}
+class FarewellServerApplicationTests extends IntegrationTestSupport {
 
 	@Test
 	void contextLoads() {
