+++ src/test/java/com/terning/farewell_server/global/common/RedisServiceIntegrationTest.java
@@ -2,18 +2,12 @@ package com.terning.farewell_server.global.common;
 
 import com.terning.farewell_server.auth.application.AuthService;
 import com.terning.farewell_server.mail.application.EmailService;
+import com.terning.farewell_server.util.IntegrationTestSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 import java.time.Duration;
 
@@ -21,23 +15,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 import static org.mockito.Mockito.mock;
 
-@SpringBootTest
-@Testcontainers
-@ActiveProfiles("test")
-class RedisServiceIntegrationTest {
+class RedisServiceIntegrationTest extends IntegrationTestSupport {
 
     @Autowired
     private RedisService redisService;
-
-    @Container
-    public static GenericContainer<?> redis = new GenericContainer<>("redis:7.0.12-alpine")
-            .withExposedPorts(6379);
-
-    @DynamicPropertySource
-    static void setRedisProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.data.redis.host", redis::getHost);
-        registry.add("spring.data.redis.port", () -> redis.getMappedPort(6379));
-    }
 
     @TestConfiguration
     static class TestConfig {
